requirements.txt
faster-whisper
transformers
torch
torchaudio
langdetect
pyannote.audio
speechbrain
textblob
scikit-learn
streamlit
pydub
